// src/app/admin-panel/store/interfaces/about.interfaces.ts
// About Interface
export interface AboutInterface {
  _id: number
  title: string
  subTitle: string
  description: string
  image: string
  createdAt: string
  __v: number
}

// Social Interface
export interface SocialInterface {
  _id: number
  facebook: string
  twitter: string
  linkedin: string
  instagram: string
  youtube: string
  createdAt: string
  __v: number
}